Add files via upload
Shorter of the last revision I printed for my video. This is not great but if you want to try it/tweak it/etc it is all yours!
This is the JBWeld I used. https://www.jbweld.com/product/plastic-bonder-syringe
This is the video link showing the CRT support repair on the TRS-80 Model III.  https://www.youtube.com/watch?v=UdRgTY8crZg